Moved leadscrew to new Elevation subsystem
