+++ roborio-src/subsystems/shooter.py
@@ -42,12 +42,7 @@ class ShooterSubsystem(Subsystem):
         # the lead screw and flywheel on the operator Xbox controller triggers.
         # The NEO 550 will be used to pull and push the note out of
         # the intake and into the flywheel.
-        self.leadscrew = FROGTalonFX(
-            constants.kLeadscrewControllerID,
-            leadscrewConfig,
-            parent_nt=f"{nt_table}",
-            motor_name="Leadscrew",
-        )
+
         self.leftFlyWheel = FROGTalonFX(
             constants.kFlyWheelControllerLeftID,
             leftFlywheelConfig,
@@ -69,24 +64,12 @@ class ShooterSubsystem(Subsystem):
         self.sequencer.setStatusFramePeriod(StatusFrameEnhanced.Status_2_Feedback0, 200)
 
         self.shooterSensor = DigitalInput(constants.kShooterSensorChannel)
-        self.shooterPositionSensor = DigitalInput(
-            constants.kShooterPositionSensorChannel
-        )
+
         self.intake = intake
 
         self._flywheelCommandedVelocity = (
             NetworkTableInstance.getDefault()
             .getFloatTopic(f"{nt_table}/FlywheelVelocity")
-            .publish()
-        )
-        self._shooterCommandedPosition = (
-            NetworkTableInstance.getDefault()
-            .getFloatTopic(f"{nt_table}/CommandedPosition")
-            .publish()
-        )
-        self._shooterActualPositionPub = (
-            NetworkTableInstance.getDefault()
-            .getFloatTopic(f"{nt_table}/ActualPosition")
             .publish()
         )
         self._sequencerCommandedPercent = (
@@ -99,13 +82,7 @@ class ShooterSubsystem(Subsystem):
             .getBooleanTopic(f"{nt_table}/NoteDetected")
             .publish()
         )
-        self._leadscrewAtPosition = (
-            NetworkTableInstance.getDefault()
-            .getBooleanTopic(f"{nt_table}/AtPosition")
-            .publish()
-        )
         self.flyWheelSpeed = 0.0
-        self.leadscrewPosition = 0.0
         self.sequencerCommandedPercent = 0.0
         self.leftFlyWheelSpeedFactor = 0.8
         self.rightFlyWheelSpeedFactor = 1.0
@@ -182,19 +159,10 @@ class ShooterSubsystem(Subsystem):
     def runSequencerCommand(self) -> Command:
         return self.runOnce(self.runSequencer).withName("RunSequencer")
 
-    def setLeadscrewCommand(self) -> Command:
-        return self.runOnce(self.moveLeadscrewToPosition).withName("SetLeadscrew")
-
     def shootCommand(self) -> Command:
         return (
             self.runFlywheelsCommand()  # run the flywheel at the commanded speed
-            .andThen(self.setLeadscrewCommand())
-            .andThen(
-                waitUntil(self.flywheelAtSpeed)
-            )  # sets the leadscrew at the commanded position
-            .andThen(
-                waitUntil(self.leadscrewAtPosition)
-            )  # wait until the leadscrew is at position
+            .andThen(waitUntil(self.flywheelAtSpeed))
             .andThen(
                 self.runSequencerCommand()
             )  # run the sequencer to move the note into the flywheel
@@ -208,71 +176,11 @@ class ShooterSubsystem(Subsystem):
             .andThen(self.stopFlywheelsCommand())
         )
 
-    def homeShooterCommand(self) -> Command:
-        return (
-            # move with positive voltage until sensor no longer reading
-            # move negative voltage until sensor reads
-            # set position
-            self.startEnd(self.runLeadscrewForward, self.stopLeadscrew)
-            .onlyWhile(self.shooterAtHome)
-            .andThen(
-                self.startEnd(self.runLeadscrewBackward, self.stopLeadscrew).until(
-                    self.shooterAtHome
-                )
-            )
-            .finallyDo(lambda interrupted: self.resetPosition(0))
-        )
-
-    def runLeadscrewForward(self):
-        self.leadscrew.set_control(VoltageOut(0.5))
-
-    def runLeadscrewBackward(self):
-        self.leadscrew.set_control(VoltageOut(-0.35))
-
-    def stopLeadscrew(self):
-        self.leadscrew.set_control(VoltageOut(0))
-
-    def runLeadscrewForwardCommand(self) -> Command:
-        return self.run(self.runLeadscrewForward)
-
-    def runLeadscrewBackwardCommand(self) -> Command:
-        return self.run(self.runLeadscrewBackward)
-
-    def resetPosition(self, position):
-        self.leadscrew.set_position(position)
-
     def stopShootingCommand(self) -> Command:
         return self.runOnce(self.stopShooting)
 
-    def setLeadscrewPosition(self, leadscrewPosition: float):
-        self.leadscrewPosition = leadscrewPosition
-
-    def getLeadscrewPosition(self) -> float:
-        return self.leadscrew.get_position().value
-
-    def moveLeadscrewToPosition(self):
-        self.leadscrew.set_control(
-            MotionMagicVoltage(position=self.leadscrewPosition, slot=1)
-        )
-
-    def leadscrewAtPosition(self) -> bool:
-        position_diff = self.leadscrewPosition - self.getLeadscrewPosition()
-        if abs(position_diff) < constants.kLeadscrewPositionTolerance:
-            return True
-        else:
-            return False
-
-    def zeroLeadscrew(self):
-        self.leadscrew.set_position(0)
-
     def noteInShooter(self) -> bool:
         return not self.shooterSensor.get()
-
-    def shooterAtHome(self) -> bool:
-        return not self.shooterPositionSensor.get()
-
-    def shooterNotAtHome(self) -> bool:
-        return self.shooterPositionSensor.get()
 
     def periodic(self) -> None:
         # self.logTelemetry()
@@ -280,17 +188,12 @@ class ShooterSubsystem(Subsystem):
 
     def publishOnSmartDashboard(self):
         SmartDashboard.putBoolean("ShooterDioSensor", self.noteInShooter())
-        SmartDashboard.putBoolean("ShooterPositionDioSensor", self.shooterAtHome())
         SmartDashboard.putBoolean("FlywheelAtSpeed", self.flywheelAtSpeed())
 
     def logTelemetry(self):
         self.leftFlyWheel.logData()
         self.rightFlyWheel.logData()
-        self.leadscrew.logData()
         self.sequencer.logData()
         self._flywheelCommandedVelocity.set(self.flyWheelSpeed)
-        self._shooterCommandedPosition.set(self.leadscrewPosition)
-        self._shooterActualPositionPub.set(self.getLeadscrewPosition())
         self._sequencerCommandedPercent.set(self.sequencerCommandedPercent)
         self._shooterSensorCurrentState.set(self.noteInShooter())
-        self._leadscrewAtPosition.set(self.leadscrewAtPosition())
